requirements.txt
absl-py==1.4.0
actionlib==1.12.1
addict==2.4.0
angles==1.9.12
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.2.0
autolab-core==1.1.1
autolab-perception==1.0.0
Babel==2.11.0
backcall==0.2.0
bleach==4.1.0
bondpy==1.8.5
cached-property==1.5.2
cachetools==4.2.4
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorlog==6.7.0
contextvars==2.4
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cv_bridge==1.13.1
cvxopt==1.3.0
cycler==0.11.0
Cython==0.29.33
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
dill==0.3.4
dynamic_reconfigure==1.6.5
entrypoints==0.4
ffmpeg-python==0.2.0
future==0.18.3
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grasp-nms==1.0.2
graspnetAPI @ file:///home/jonesubu18/PycharmProjects/pythonProject/graspnetAPI
grpcio==1.48.2
h5py==3.1.0
idna==3.4
image_geometry==1.13.1
imageio==2.15.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
interactive-markers==1.11.5
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.3
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
joint_state_publisher==1.12.15
joint_state_publisher_gui==1.12.15
json5==0.9.11
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.2
kdl-parser-py==1.13.3
kiwisolver==1.3.1
knn-pytorch==0.1
laser_geometry==1.6.7
Markdown==3.3.7
MarkupSafe==2.0.1
mat73==0.60
matplotlib==3.3.4
message_filters==1.14.13
mistune==0.8.4
moveit-core==1.0.11
moveit_commander==1.0.11
moveit_ros_planning_interface==1.0.11
moveit_ros_visualization==1.0.11
multiprocess==0.70.12.2
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
networkx==2.5.1
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.2
open3d==0.15.2
opencv-python==4.7.0.72
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pointnet2==0.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrealsense2==2.53.1.4623
pyrsistent==0.18.0
pyserial==3.5
python-dateutil==2.8.2
python_qt_binding==0.4.4
pytz==2022.7.1
PyWavefront==1.3.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==25.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.27.1
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_joint_trajectory_controller==0.17.3
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
rviz==1.13.29
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
sensor-msgs==1.12.8
setproctitle==1.2.3
six==1.16.0
sklearn==0.0.post1
smach==2.0.1
smach_ros==2.0.1
smclib==1.8.5
sniffio==1.2.0
srdfdom==0.5.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.6.0
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
threadpoolctl==3.1.0
tifffile==2020.9.3
tk==0.1.0
tomlkit==0.11.6
topic_tools==1.14.13
torch==1.6.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
transforms3d==0.3.1
trimesh==3.20.2
typing_extensions==4.1.1
urdfdom-py==0.4.6
urllib3==1.26.15
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.2
xacro==1.13.18
zipp==3.6.0
